Team dropdowns and corrected json errors
A team list was implemetented that allows the scorekeeper to choose the
team that is playing from a dropdown. The dropdown is auto-filled using
a text file with a team on each line. We corrected some point value
errors and point maximum errors in the json file for the input elements.